Close #29 - Rename package
