+++ core/src/main/scala/maven2sbt/core/Common.scala
@@ -1,4 +1,4 @@
-package kevinlee.maven2sbt
+package maven2sbt.core
 
 import scala.util.matching.Regex
