+++ core/src/main/scala/maven2sbt/core/Exclusion.scala
@@ -1,4 +1,4 @@
-package kevinlee.maven2sbt
+package maven2sbt.core
 
 import Common._
