+++ core/src/main/scala/maven2sbt/core/ProjectInfo.scala
@@ -1,4 +1,4 @@
-package kevinlee.maven2sbt
+package maven2sbt.core
 
 import scala.language.postfixOps
 import scala.xml.Elem
