+++ core/src/main/scala/maven2sbt/core/data.scala
@@ -1,4 +1,4 @@
-package kevinlee.maven2sbt
+package maven2sbt.core
 
 /**
   * @author Kevin Lee
